Move classloader registry out of worker services as its no longer needed
