+++ subprojects/core/src/main/java/org/gradle/initialization/DefaultClassLoaderRegistry.java
@@ -18,14 +18,12 @@ package org.gradle.initialization;
 
 import org.gradle.api.internal.ClassPathRegistry;
 import org.gradle.internal.classloader.FilteringClassLoader;
-import org.gradle.internal.classpath.ClassPath;
 import org.gradle.internal.reflect.Instantiator;
 
 public class DefaultClassLoaderRegistry implements ClassLoaderRegistry {
     private final ClassLoader apiOnlyClassLoader;
     private final ClassLoader apiAndPluginsClassLoader;
     private final ClassLoader pluginsClassLoader;
-    private final ClassLoader workerPluginsClassLoader;
     private final FilteringClassLoader.Spec gradleApiSpec;
     private final MixInLegacyTypesClassLoader.Spec workerExtensionSpec;
     private final Instantiator instantiator;
@@ -35,10 +33,8 @@ public class DefaultClassLoaderRegistry implements ClassLoaderRegistry {
         ClassLoader runtimeClassLoader = getClass().getClassLoader();
         this.apiOnlyClassLoader = restrictToGradleApi(runtimeClassLoader);
         this.pluginsClassLoader = new MixInLegacyTypesClassLoader(runtimeClassLoader, classPathRegistry.getClassPath("GRADLE_EXTENSIONS"), legacyTypesSupport);
-        ClassPath workerExtensionClassPath = classPathRegistry.getClassPath("GRADLE_WORKER_EXTENSIONS");
-        this.workerPluginsClassLoader = new MixInLegacyTypesClassLoader(runtimeClassLoader, workerExtensionClassPath, legacyTypesSupport);
         this.gradleApiSpec = apiSpecFor(pluginsClassLoader);
-        this.workerExtensionSpec = new MixInLegacyTypesClassLoader.Spec("legacy-mixin-loader", workerExtensionClassPath.getAsURLs());
+        this.workerExtensionSpec = new MixInLegacyTypesClassLoader.Spec("legacy-mixin-loader", classPathRegistry.getClassPath("GRADLE_WORKER_EXTENSIONS").getAsURLs());
         this.apiAndPluginsClassLoader = restrictTo(gradleApiSpec, pluginsClassLoader);
     }
 
@@ -86,11 +82,6 @@ public class DefaultClassLoaderRegistry implements ClassLoaderRegistry {
     @Override
     public ClassLoader getGradleCoreApiClassLoader() {
         return apiOnlyClassLoader;
-    }
-
-    @Override
-    public ClassLoader getWorkerPluginsClassLoader() {
-        return workerPluginsClassLoader;
     }
 
     @Override
