+++ subprojects/core/src/main/java/org/gradle/initialization/FlatClassLoaderRegistry.java
@@ -47,11 +47,6 @@ public class FlatClassLoaderRegistry implements ClassLoaderRegistry {
     }
 
     @Override
-    public ClassLoader getWorkerPluginsClassLoader() {
-        return classLoader;
-    }
-
-    @Override
     public FilteringClassLoader.Spec getGradleApiFilterSpec() {
         throw new UnsupportedOperationException();
     }
